healthbridge-api/requirements.txt
# HealthBridge AI — Python Dependencies

# ─── Web Framework ───
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ─── Database ───
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2

# ─── Redis ───
redis[hiredis]==5.1.0

# ─── Validation & Settings ───
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# ─── Auth ───
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ─── LLM SDKs ───
google-generativeai==0.8.0
groq==0.11.0

# ─── PDF Generation ───
reportlab==4.2.0

# ─── HTTP Client ───
httpx==0.27.0

# ─── Task Queue ───
celery[redis]==5.4.0

# ─── Logging ───
python-json-logger==2.0.7

# ─── CORS & Middleware ───
starlette==0.38.0
